requirements.txt
absl-py==1.4.0
affine==2.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
Automat==22.10.0
autopep8==2.0.2
backcall==0.2.0
bbox-visualizer==0.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
clearml==1.11.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cmake==3.26.3
colorama==0.4.6
Comet==3.1.0
comm==0.1.3
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
etils==1.2.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
flatbuffers==23.5.8
fonttools==4.39.3
fqdn==1.5.1
furl==2.1.3
gast==0.4.0
GDAL==3.6.4
gin-config==0.5.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
h5py==3.8.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
imageio==2.28.1
immutabledict==2.2.4
importlib-resources==5.12.0
incremental==22.10.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jax==0.4.9
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-bbox-widget==0.5.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.5.13
keras==2.12.0
keras-cv==0.5.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
lit==16.0.3
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.1.0
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.Open the command palette91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
portalocker==2.7.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==4.23.0
protos==0.1.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.4.0
pylabel==0.1.50
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
rasterio==1.3.6
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
sacrebleu==2.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.22.2
seqeval==1.2.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-addons==0.20.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorflow-model-optimization==0.7.4
tensorflow-text==2.12.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
tf-models-official==2.12.0
tf-slim==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
toml==0.10.2
torch==2.0.1
torch-tb-profiler==0.4.1
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
Twisted==22.10.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.101
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wrapt==1.14.1
zipp==3.15.0
zope.interface==6.0
